feat(studio): add Tailwind CSS support
Add Tailwind CSS to the project by installing the required dependencies and
configuring the necessary files. The following dependencies were added:
- autoprefixer
- postcss
- sass
- tailwindcss

The following files were created:
- postcss.config.js
- tailwind.config.js

The `index.scss` file was created to import the Tailwind CSS styles. The
`tailwind.config.js` file was configured to specify the content and theme of the
project. The `postcss.config.js` file was created to configure the PostCSS
plugins.

The `router/index.tsx` file was modified to add a sample Tailwind CSS class to
the `Hello Agnost` text.