feat(Avatar): add @radix-ui/react-avatar dependency and Avatar component
This commit adds the @radix-ui/react-avatar dependency and a new Avatar component to the project. The Avatar component is a wrapper around the AvatarPrimitive.Root component from the @radix-ui/react-avatar package. It accepts a size prop that can be one of 'xs', 'sm', 'md', 'lg', or 'xl'. The Avatar component also exports AvatarImage and AvatarFallback components that are wrappers around AvatarPrimitive.Image and AvatarPrimitive.Fallback components respectively. The AvatarFallback component accepts a name prop that is used to generate the fallback avatar image. The styles for the Avatar component are defined in the avatar.scss file.