feat(VerificationCodeInput): add VerificationCodeInput component
Add VerificationCodeInput component to the project. This component is used to input verification codes. It is implemented using the react-verification-input library. The component has a length of 6 and only accepts numbers as valid characters. It also has an error state that can be set to true to indicate an error. The component is styled using the verificationCodeInput.scss file.

Update the CompleteAccountSetupVerifyEmail page to use the VerificationCodeInput component instead of the placeholder text. The component is used to input the verification code sent to the user's email. The onChange and onComplete props are used to handle the input and completion of the verification code.