feat(Avatar.tsx, avatar.scss): add support for differentiating user and organization avatars
Add a new prop `isUserAvatar` to the `AvatarFallbackProps` interface to differentiate between user and organization avatars. In the `AvatarFallback` component, add a conditional class name based on the value of `isUserAvatar`. In the `avatar.scss` file, add two new classes, `avatar-fallback-user` and `avatar-fallback-org`, to differentiate the styles for user and organization avatars.