feat(OrganizationService.ts, organizationStore.ts): add createOrganization method to OrganizationService and organizationStore.
The createOrganization method creates a new organization with the given name and returns the created organization. The organization is also added to the list of organizations in the store. If the creation is successful, the onSuccess callback is called. If there is an error, the onError callback is called and the error is thrown.

Also, the return type of getAllOrganizationsByUser method in OrganizationService is changed to Promise<Organization[]> to reflect the actual return type.

Finally, a loading state is added to the organizationStore to indicate when the getAllOrganizationByUser method is fetching data.