feat(CopyInput): add CopyInput component
The CopyInput component is a custom input component that allows users to copy the input value to the clipboard. It consists of an input field and a copy button. When the copy button is clicked, the input value is copied to the clipboard using the Clipboard API. The component is styled using SCSS.