feat(Command): add Command component
The Command component is a custom implementation of the cmdk library's
Command component. It provides a searchable list of commands that can be
executed by the user. The component is composed of several sub-components
such as CommandDialog, CommandInput, CommandList, CommandEmpty,
CommandGroup, CommandItem, CommandShortcut, and CommandSeparator. The
styling for the component is defined in the command.scss file. The
component is exported from the index.ts file.