feat(Dialog): add Dialog component with its subcomponents
This commit adds a new Dialog component with its subcomponents to the
project. The Dialog component is a reusable component that can be used
to display a modal dialog box. The Dialog component is built using the
Radix UI library. The Dialog component has the following subcomponents:
- DialogTrigger
- DialogContent
- DialogHeader
- DialogFooter
- DialogTitle
- DialogDescription

This commit also adds a SCSS file for the Dialog component. The SCSS
file contains the styles for the Dialog component and its subcomponents.