feat(OrganizationDropdown.tsx): add i18n support for organization dropdown
This commit adds i18n support for the organization dropdown component. It
uses the `useTranslation` hook from `react-i18next` to translate the
strings in the component. The following strings were added to the
`organization.json` file:

- `organization.select`
- `organization.empty`
- `organization.leave`
- `organization.leave.desc`
- `organization.leave.success`
- `organization.leave.success.desc`

The `organization-dropdown-content` class was added to the
`PopoverContent` component to allow for custom styling of the dropdown
content. The `organization-label` class was added to the `div` element
that contains the organization avatar and name.