feat(ApplicationCardSkeleton): add skeleton component for application card
This commit adds a new skeleton component for the application card. The
skeleton component is used to display a loading state while the actual
data is being fetched. The component is implemented using the
ApplicationCardSkeleton function which returns a JSX element. The
component is styled using the skeleton.scss file which contains the
styles for the skeleton component. The component is exported from the
index.ts file so that it can be used in other parts of the application.